# api-gateway/internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Backend API
  version: 1.0.0
  description: Public API served through the API Gateway. Internal service endpoints are intentionally excluded.
servers:
  - url: /
security: []
...
paths:
...
  /users/signup:
    post:
      tags: [Users]
      summary: Register a manual user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /users/me:
    get:
      tags: [Users]
      summary: Get own profile
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/OK"
    put:
      tags: [Users]
      summary: Partially update own profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          $ref: "#/components/responses/OK"
    delete:
      tags: [Users]
      summary: Soft-delete own account
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /users/me/password:
    put:
      tags: [Users]
      summary: Change own password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          $ref: "#/components/responses/OK"
...
  /users:
    get:
      tags: [Admin]
      summary: List users
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /users/{id}:
    get:
      tags: [Admin]
      summary: Get any user
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          $ref: "#/components/responses/OK"
    delete:
      tags: [Admin]
      summary: Soft-delete a user
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /users/{id}/ban:
    put:
      tags: [Admin]
      summary: Ban or unban a user
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_banned]
              properties:
                is_banned:
                  type: boolean
                reason:
                  type: string
                  maxLength: 1000
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /users/{id}/role:
    put:
      tags: [Admin]
      summary: Change user role
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [user, admin, superadmin]
      responses:
        "200":
          $ref: "#/components/responses/OK"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    OK:
      description: Success
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Success"
...
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Rate limited
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Success:
      type: object
      required: [success, meta]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          additionalProperties: true
        meta:
          $ref: "#/components/schemas/Meta"
    ErrorResponse:
      type: object
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
            retry_after_seconds:
              type: integer
        meta:
          $ref: "#/components/schemas/Meta"
    Meta:
      type: object
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
...
    SignupRequest:
      type: object
      required: [full_name, username, email, phone, password, gouvernorat_id]
      properties:
        full_name:
          type: string
          minLength: 2
          maxLength: 150
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
          maxLength: 255
        phone:
          type: string
          maxLength: 20
          example: "+21612345678"
        password:
          type: string
          minLength: 8
          maxLength: 128
          format: password
        gouvernorat_id:
          type: integer
          minimum: 1
          maximum: 24
...
    UpdateProfileRequest:
      type: object
      properties:
        full_name:
          type: string
          minLength: 2
          maxLength: 150
        username:
          type: string
          minLength: 3
          maxLength: 50
        phone:
          type: string
          maxLength: 20
        gouvernorat_id:
          type: integer
          minimum: 1
          maximum: 24
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          maxLength: 128
          format: password
        new_password:
          type: string
          minLength: 8
          maxLength: 128
          format: password
...